+++ CHANGELOG.md
@@ -10,6 +10,18 @@ All notable changes to `iex-cloud-sdk` will be documented in this file
 
 - Social Sentiment
 
+##### Investors Exchange Data
+
+- DEEP Operational Halt Status
+- DEEP Official Price
+- DEEP Short Sale Price Test Status
+- DEEP Trade Break
+- DEEP Trading Status
+
+##### Reference Data
+
+- Search
+
 ## v0.1.0-beta.5 - 2019-10-02
 
 ### Added